chore: implement the `pid` crate
