+++ kernel/Cargo.toml
@@ -43,6 +43,7 @@ arrayvec = { version = "0.7.1", default-features = false }
 posix = { path = "../libs/posix" }
 num-derive = { version = "0.3.3", default-features = false }
 num-traits = { version = "0.2.14", default-features = false }
+pid = { path = "../libs/pid" }
 
 [build-dependencies]
 cc = "1.0.70"
